+++ flink-core/src/main/java/org/apache/flink/core/mailbox/MailboxExecutor.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.api.operators;
+package org.apache.flink.core.mailbox;
 
 import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.runtime.concurrent.FutureTaskWithException;
