+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/operators/YieldingOperatorFactory.java
@@ -18,6 +18,7 @@
 package org.apache.flink.streaming.api.operators;
 
 import org.apache.flink.annotation.Experimental;
+import org.apache.flink.core.mailbox.MailboxExecutor;
 
 /**
  * An operator that needs access to the {@link MailboxExecutor} to yield to downstream operators
