+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/ProcessingTimeServiceFactory.java
@@ -18,7 +18,7 @@
 package org.apache.flink.streaming.runtime.tasks;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.streaming.api.operators.MailboxExecutor;
+import org.apache.flink.core.mailbox.MailboxExecutor;
 
 /**
  * A factory for creating processing time services with a given {@link MailboxExecutor}. The factory
