+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/mailbox/MailboxExecutorFactory.java
@@ -18,7 +18,7 @@
 package org.apache.flink.streaming.runtime.tasks.mailbox;
 
 import org.apache.flink.annotation.Internal;
-import org.apache.flink.streaming.api.operators.MailboxExecutor;
+import org.apache.flink.core.mailbox.MailboxExecutor;
 
 /**
  * A factory for creating mailbox executors with a given priority. The factory is usually bound to a
