+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/api/operators/SyncMailboxExecutor.java
@@ -17,6 +17,7 @@
 
 package org.apache.flink.streaming.api.operators;
 
+import org.apache.flink.core.mailbox.MailboxExecutor;
 import org.apache.flink.util.FlinkRuntimeException;
 import org.apache.flink.util.function.ThrowingRunnable;
